Добавлен функционал магазина.

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -3,13 +3,13 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Helpers\Admin\DbSort;
+use App\Helpers\Admin\Img;
+use App\Models\Category;
+use App\Models\Main;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\DB;
-use Illuminate\Support\Facades\Schema;
 use Illuminate\Support\Str;
-use App\Models\Dummy;
 
-class DummyController extends AppController
+class CategoryController extends AppController
 {
     public function __construct(Request $request)
     {
@@ -22,24 +22,7 @@ class DummyController extends AppController
         $route = $this->route = $request->segment(2);
         $view = $this->view = Str::snake($this->class);
 
-        // Связанная таблица, должен быть метод в моделе с названием таблицы
-        $belongTable = $this->belongTable = '';
-
-        // Связанные таблицы, а также в моделе должен быть метод с название таблицы, реализующий связь
-        $relatedTables = $this->relatedTables = [
-
-            // Категории
-            //'categories',
-        ];
-
-        // Связанные таблицы, которые нельзя удалить, если есть связанные элементы, а также в моделе должен быть метод с название таблицы, реализующий связь
-        $relatedDelete = $this->relatedDelete = [
-
-            // Формы
-            //'forms',
-        ];
-
-        view()->share(compact('class', 'c','model', 'table', 'route', 'view', 'belongTable', 'relatedTables', 'relatedDelete'));
+        view()->share(compact('class', 'c','model', 'table', 'route', 'view'));
     }
 
     /**
@@ -47,11 +30,15 @@ class DummyController extends AppController
      *
      * @return \Illuminate\Http\Response
      */
-    public function index(Request $request)
+    public function index()
     {
         // Поиск. Массив гет ключей для поиска
         $queryArr = [
             'title',
+            'slug',
+            'status',
+            'sort',
+            'parent_id',
             'id',
         ];
 
@@ -65,7 +52,12 @@ class DummyController extends AppController
 
         // Передать поля для вывода, значение l - с переводом, t - дата
         $thead = [
+            'img' => 'img',
             'title' => null,
+            'slug' => null,
+            'status' => 'l',
+            'sort' => null,
+            'parent_id' => null,
             'id' => null,
         ];
 
@@ -95,14 +87,28 @@ class DummyController extends AppController
      */
     public function store(Request $request)
     {
+        $magesExt = implode(',', config('admin.acceptedImagesExt') ?? []);
         $rules = [
-            'title' => "required|string|unique:{$this->table}|max:250",
+            'title' => 'required|string|max:250',
+            'slug' => "required|string|unique:{$this->table}|max:250",
+            'img' => "nullable|mimes:{$magesExt}",
         ];
         $request->validate($rules);
         $data = $request->all();
 
+        if ($request->hasFile('img')) {
+
+            // Обработка картинки
+            $data['img'] = Img::upload($request, $this->class);
+
+        } else {
+
+            // Если нет картинки
+            $data['img'] = config("admin.img{$this->class}Default");
+        }
+
         // Создаём экземкляр модели
-        $values = new Dummy();
+        $values = new Category();
 
         // Заполняем модель новыми данными
         $values->fill($data);
@@ -140,24 +146,18 @@ class DummyController extends AppController
     {
         $values = $this->model::findOrFail($id);
 
-        // Получаем данные связанных таблиц
-        $related = [];
-        if (!empty($this->relatedTables)) {
-            foreach ($this->relatedTables as $relatedTable) {
-                if (Schema::hasTable($relatedTable)) {
-                    $related[$relatedTable] = DB::table($relatedTable)
-                        ->whereNull('deleted_at')
-                        ->pluck('title', 'id');
-                }
-            }
-        }
+        // Записать в реестр parent_id, для построения дерева
+        Main::set('parent_id', $values->parent_id);
+
+        // Получаем все элементы в массив, где ключи id
+        $all = $this->model::get()->keyBy('id')->toArray();
 
         // Элементы связанные
         $valuesBelong = $values->{$this->table};
 
         $f = __FUNCTION__;
         $title = __("a.{$f}");
-        return view("{$this->viewPath}.{$this->view}.{$this->template}", compact('title', 'values', 'related', 'valuesBelong'));
+        return view("{$this->viewPath}.{$this->view}.{$this->template}", compact('title', 'values', 'all', 'valuesBelong'));
     }
 
     /**
@@ -172,20 +172,31 @@ class DummyController extends AppController
         // Получаем элемент по id, если нет - будет ошибка
         $values = $this->model::findOrFail($id);
 
+        $magesExt = implode(',', config('admin.acceptedImagesExt') ?? []);
         $rules = [
-            'title' => "required|string|unique:{$this->table},title,{$id}|max:250",
+            'title' => 'required|string|max:250',
+            'slug' => "required|string|unique:{$this->table},slug,{$id}|max:250",
+            'parent_id' => 'nullable|integer',
+            'img' => "nullable|mimes:{$magesExt}",
         ];
         $request->validate($rules);
         $data = $request->all();
 
+        if ($request->hasFile('img')) {
 
-        // Сохраняем связи
-        if (!empty($this->relatedTables)) {
-            foreach ($this->relatedTables as $relatedTable) {
-                $values->$relatedTable()->sync($request->$relatedTable);
-            }
+            // Обработка картинки
+            $data['img'] = Img::upload($request, $this->class, $values->img);
+
+        } else {
+
+            // Если нет картинки
+            $data['img'] = $values->img;
         }
 
+        // parent_id не должны быть равно id
+        if ($values->parent_id == $values->id) {
+            $values->parent_id = '0';
+        }
 
         // Заполняем модель новыми данными
         $values->fill($data);
@@ -213,25 +224,15 @@ class DummyController extends AppController
         // Получаем элемент по id, если нет - будет ошибка
         $values = $this->model::findOrFail($id);
 
-
-        // Если есть связи, то вернём ошибку
-        if (!empty($this->relatedDelete)) {
-            foreach ($this->relatedDelete as $relatedTable) {
-                if ($values->$relatedTable->count()) {
-                    return redirect()
-                        ->route("admin.{$this->route}.edit", $id)
-                        ->with('error', __('s.remove_not_possible') . ', ' . __('s.there_are_nested') . __('a.id'));
-                }
-            }
+        // Если есть потомки, то ошибка
+        if ($values->parents->isNotEmpty()) {
+            return redirect()
+                ->route("admin.{$this->route}.edit", $id)
+                ->with('error', __('s.remove_not_possible') . ', ' . __('s.there_are_nested') . __('a.id'));
         }
 
-
-        // Удаляем связанные элементы
-        if (!empty($this->relatedTables)) {
-            foreach ($this->relatedTables as $relatedTable) {
-                $values->$relatedTable()->sync([]);
-            }
-        }
+        // Удалить картинку, кроме картинки по-умолчанию
+        Img::deleteImg($values->img, config("admin.img{$this->class}Default"));
 
         // Удаляем элемент
         $values->delete();
